Atttempting to use FontSquirrel's font kit to make fonts work on Firefox
The OFT file seems to work fine with Safari and Chrome (on Mac OS X),
but Mozilla Firefox not only does not show the font as specified,
it fails to show any of the fallback fonts either. I think this may
be a bug or something mysterious. SO I have tried replacing League Gothic
with the Font Squirrel version (basically the addition of WOFF and
EOT files for browsers that prefer them). Still no joy with Firefox.